# contract/openapi.yaml
openapi: 3.0.3
info:
  title: AIDetector API
  version: 1.0.0
  description: V1.0 active contract baseline for auth, detection, history, quota, admin, report, key, and team APIs.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8000
    description: Local
...
paths:
...
  /api/v1/admin/status:
    get:
      tags: [admin]
      summary: Verify admin access
      operationId: getAdminStatus
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Admin access confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStatusResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/overview:
    get:
      tags: [admin]
      summary: Get admin overview metrics
      operationId: getAdminOverview
      security:
        - BearerAuth: []
      parameters:
        - name: preset
          in: query
          schema:
            type: string
            enum: [today, week, month, quarter, year]
            default: week
      responses:
        '200':
          description: Overview data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOverviewResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/users:
    get:
      tags: [admin]
      summary: List users for admin
      operationId: listAdminUsers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: systemRole
          in: query
          schema:
            $ref: '#/components/schemas/UserSystemRole'
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: planTier
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, email, creditsRemaining]
            default: createdAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/users/{userId}:
    get:
      tags: [admin]
      summary: Get user detail for admin
      operationId: getAdminUser
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetailResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [admin]
      summary: Update admin-managed user fields
      operationId: updateAdminUser
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserUpdateRequest'
      responses:
        '200':
          description: Updated user detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetailResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/users/{userId}/credits:
    post:
      tags: [admin]
      summary: Adjust user credits
      operationId: adjustAdminUserCredits
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserCreditsAdjustRequest'
      responses:
        '200':
          description: Updated credit snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserCreditsAdjustResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/detections:
    get:
      tags: [admin]
      summary: List detections for admin
      operationId: listAdminDetections
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
        - name: actorType
          in: query
          schema:
            type: string
            enum: [user, guest]
        - name: label
          in: query
          schema:
            type: string
            enum: [ai, mixed, human]
        - name: function
          in: query
          schema:
            $ref: '#/components/schemas/DetectionFunction'
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, score, charsUsed]
            default: createdAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Detection list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDetectionListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/detections/{detectionId}:
    get:
      tags: [admin]
      summary: Get detection detail for admin
      operationId: getAdminDetection
      security:
        - BearerAuth: []
      parameters:
        - name: detectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detection detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDetectionDetailResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Detection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [admin]
      summary: Delete detection record
      operationId: deleteAdminDetection
      security:
        - BearerAuth: []
      parameters:
        - name: detectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Detection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - detail
      properties:
        code:
          oneOf:
            - type: integer
            - type: string
        message:
          type: string
        detail:
          nullable: true
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: string
...
    UserSystemRole:
      type: string
      enum: [VISITOR, INDIVIDUAL, TEAM_ADMIN, SYS_ADMIN]
    DetectionFunction:
      type: string
      enum: [scan]
...
    UserProfile:
      type: object
      properties:
        firstName:
          type: string
          nullable: true
        surname:
          type: string
          nullable: true
        organization:
          type: string
          nullable: true
        jobRole:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
...
    HistorySummary:
      type: object
      required:
        - ai
        - mixed
        - human
      properties:
        ai:
          type: integer
          minimum: 0
          maximum: 100
        mixed:
          type: integer
          minimum: 0
          maximum: 100
        human:
          type: integer
          minimum: 0
          maximum: 100
    HistorySentence:
      type: object
      required:
        - id
        - text
        - raw
        - type
        - probability
        - score
        - reason
        - suggestion
      properties:
        id:
          type: string
        text:
          type: string
        raw:
          type: string
        startParagraph:
          type: integer
          nullable: true
          minimum: 1
        endParagraph:
          type: integer
          nullable: true
          minimum: 1
        type:
          type: string
          enum: [ai, mixed, human]
        probability:
          type: number
          format: float
          minimum: 0
          maximum: 1
        score:
          type: integer
          minimum: 0
          maximum: 100
        reason:
          type: string
        suggestion:
          type: string
...
    HistoryAnalysis:
      type: object
      description: >
        Analysis payload used by current scan/history flows. The translation, polish, and citations
        fields are compatibility placeholders and are not populated by active product features.
      required:
        - summary
        - sentences
        - translation
        - polish
        - citations
        - aiLikelyCount
        - highlightedHtml
      properties:
        summary:
          $ref: '#/components/schemas/HistorySummary'
        sentences:
          type: array
          items:
            $ref: '#/components/schemas/HistorySentence'
        translation:
          type: string
          description: Reserved compatibility field. Currently returned as an empty string.
        polish:
          type: string
          description: Reserved compatibility field. Currently returned as an empty string.
        citations:
          type: array
          description: Reserved compatibility field. Currently returned as an empty array.
          items:
            $ref: '#/components/schemas/HistoryCitation'
        aiLikelyCount:
          type: integer
          minimum: 0
        highlightedHtml:
          type: string
...
    AdminStatusResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    AdminOverviewPreset:
      type: string
      enum: [today, week, month, quarter, year]
    AdminOverviewGranularity:
      type: string
      enum: [hour, day, week, month]
    AdminOverviewPeriod:
      type: object
      required:
        - preset
        - granularity
        - startAt
        - endAt
      properties:
        preset:
          $ref: '#/components/schemas/AdminOverviewPreset'
        granularity:
          $ref: '#/components/schemas/AdminOverviewGranularity'
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
    AdminOverviewSummary:
      type: object
      required:
        - totalUsers
        - activeUsers
        - newUsers
        - detections
        - charsUsed
      properties:
        totalUsers:
          type: integer
        activeUsers:
          type: integer
        newUsers:
          type: integer
        detections:
          type: integer
        charsUsed:
          type: integer
    AdminOverviewSeriesItem:
      type: object
      required:
        - bucketStart
        - bucketLabel
        - newUsers
        - detections
        - charsUsed
      properties:
        bucketStart:
          type: string
          format: date-time
        bucketLabel:
          type: string
        newUsers:
          type: integer
        detections:
          type: integer
        charsUsed:
          type: integer
    AdminRecentUserItem:
      type: object
      required:
        - id
        - email
        - name
        - systemRole
        - isActive
        - createdAt
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
        systemRole:
          $ref: '#/components/schemas/UserSystemRole'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AdminRecentDetectionItem:
      type: object
      required:
        - id
        - actorType
        - actorId
        - label
        - score
        - charsUsed
        - createdAt
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        userEmail:
          type: string
          format: email
          nullable: true
        userName:
          type: string
          nullable: true
        actorType:
          type: string
          enum: [user, guest]
        actorId:
          type: string
        label:
          type: string
          enum: [ai, mixed, human]
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        charsUsed:
          type: integer
        createdAt:
          type: string
          format: date-time
    AdminOverviewResponse:
      type: object
      required:
        - period
        - summary
        - series
        - recentUsers
        - recentDetections
      properties:
        period:
          $ref: '#/components/schemas/AdminOverviewPeriod'
        summary:
          $ref: '#/components/schemas/AdminOverviewSummary'
        series:
          type: array
          items:
            $ref: '#/components/schemas/AdminOverviewSeriesItem'
        recentUsers:
          type: array
          items:
            $ref: '#/components/schemas/AdminRecentUserItem'
        recentDetections:
          type: array
          items:
            $ref: '#/components/schemas/AdminRecentDetectionItem'
    AdminUserListItem:
      type: object
      required:
        - id
        - email
        - name
        - systemRole
        - planTier
        - isActive
        - creditsTotal
        - creditsUsed
        - creditsRemaining
        - createdAt
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
        systemRole:
          $ref: '#/components/schemas/UserSystemRole'
        planTier:
          type: string
        isActive:
          type: boolean
        creditsTotal:
          type: integer
        creditsUsed:
          type: integer
        creditsRemaining:
          type: integer
        createdAt:
          type: string
          format: date-time
    AdminUserListResponse:
      type: object
      required:
        - items
        - page
        - pageSize
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserListItem'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    AdminDetectionMini:
      type: object
      required:
        - id
        - label
        - score
        - charsUsed
        - createdAt
      properties:
        id:
          type: integer
        label:
          type: string
          enum: [ai, mixed, human]
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        charsUsed:
          type: integer
        createdAt:
          type: string
          format: date-time
    AdminUserDetailResponse:
      allOf:
        - $ref: '#/components/schemas/AdminUserListItem'
        - type: object
          properties:
            profile:
              $ref: '#/components/schemas/UserProfile'
            recentDetections:
              type: array
              items:
                $ref: '#/components/schemas/AdminDetectionMini'
    AdminUserUpdateRequest:
      type: object
      properties:
        systemRole:
          $ref: '#/components/schemas/UserSystemRole'
        planTier:
          type: string
        isActive:
          type: boolean
    AdminUserCreditsAdjustRequest:
      type: object
      required:
        - delta
        - reason
      properties:
        delta:
          type: integer
          description: Positive to grant credits, negative to deduct credits
        reason:
          type: string
          minLength: 1
          maxLength: 255
    AdminUserCreditsAdjustResponse:
      type: object
      required:
        - userId
        - creditsTotal
        - creditsUsed
        - creditsRemaining
      properties:
        userId:
          type: integer
        creditsTotal:
          type: integer
        creditsUsed:
          type: integer
        creditsRemaining:
          type: integer
    AdminDetectionListItem:
      type: object
      required:
        - id
        - actorType
        - actorId
        - label
        - score
        - charsUsed
        - functionsUsed
        - createdAt
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        userEmail:
          type: string
          format: email
          nullable: true
        userName:
          type: string
          nullable: true
        actorType:
          type: string
          enum: [user, guest]
        actorId:
          type: string
        label:
          type: string
          enum: [ai, mixed, human]
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        charsUsed:
          type: integer
        functionsUsed:
          type: array
          items:
            $ref: '#/components/schemas/DetectionFunction'
        createdAt:
          type: string
          format: date-time
    AdminDetectionListResponse:
      type: object
      required:
        - items
        - page
        - pageSize
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminDetectionListItem'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    AdminDetectionDetailResponse:
      allOf:
        - $ref: '#/components/schemas/AdminDetectionListItem'
        - type: object
          properties:
            title:
              type: string
              nullable: true
            inputText:
              type: string
            editorHtml:
              type: string
              nullable: true
            metaJson:
              type: object
              nullable: true
              additionalProperties: true
            analysis:
              $ref: '#/components/schemas/HistoryAnalysis'
              nullable: true